add item

+++ src/Objects/Item.java
@@ -70,7 +70,7 @@ public class Item {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(float price) {
         this.price = price;
     }
